requirements.txt
torch
torchvision
scipy
scikit-image
numpy
imutils
psd-tools
opencv-python
mediapipe
cmake
dlib
face_cropper
Pillow
tensorboardX
pandas
huggingface_hub